fixing structure and assets
