+++ flutter/beagle/assets/js/beagle.js
