+++ flutter/beagle/javascript-bridge/package.json
