+++ flutter/beagle/javascript-bridge/src/global.d.ts
@@ -1,0 +1,8 @@
+type FlutterJSChannel = (
+    'httpClient.request'
+    | 'beagleView.update'
+    | 'action'
+  )
+  
+  declare function sendMessage(channel: FlutterJSChannel, message: String)
+
