+++ flutter/beagle/javascript-bridge/src/index.ts
