+++ flutter/beagle/javascript-bridge/src/utils/promise.ts
