+++ flutter/beagle/javascript-bridge/webpack.config.js
