+++ flutter/beagle/lib/bridge_impl/beagle_service_js.dart
@@ -15,19 +15,19 @@
  *  limitations under the License.
  */
 
-import 'package:beagle_core/bridge_impl/beagle_js_engine.dart';
-import 'package:beagle_core/bridge_impl/beagle_view_js.dart';
-import 'package:beagle_core/default/default_action_handler.dart';
-import 'package:beagle_core/default/default_http_client.dart';
-import 'package:beagle_core/interface/action_handler.dart';
-import 'package:beagle_core/interface/beagle_service.dart';
-import 'package:beagle_core/interface/beagle_view.dart';
-import 'package:beagle_core/interface/component_builder.dart';
-import 'package:beagle_core/interface/http_client.dart';
-import 'package:beagle_core/interface/storage.dart';
-import 'package:beagle_core/model/network_options.dart';
-import 'package:beagle_core/model/network_strategy.dart';
-import 'package:beagle_core/model/request.dart';
+import 'package:beagle/bridge_impl/beagle_js_engine.dart';
+import 'package:beagle/bridge_impl/beagle_view_js.dart';
+import 'package:beagle/default/default_action_handler.dart';
+import 'package:beagle/default/default_http_client.dart';
+import 'package:beagle/interface/action_handler.dart';
+import 'package:beagle/interface/beagle_service.dart';
+import 'package:beagle/interface/beagle_view.dart';
+import 'package:beagle/interface/component_builder.dart';
+import 'package:beagle/interface/http_client.dart';
+import 'package:beagle/interface/storage.dart';
+import 'package:beagle/model/network_options.dart';
+import 'package:beagle/model/network_strategy.dart';
+import 'package:beagle/model/request.dart';
 
 class BeagleServiceJS implements BeagleService {
   BeagleServiceJS({
