+++ flutter/beagle/lib/interface/http_client.dart
@@ -1,13 +1,13 @@
 /*
- *  
+ *
  *  Copyright 2020 ZUP IT SERVICOS EM TECNOLOGIA E INOVACAO SA
- *  
+ *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
  *  You may obtain a copy of the License at
- *  
+ *
  *        http://www.apache.org/licenses/LICENSE-2.0
- *  
+ *
  *  Unless required by applicable law or agreed to in writing, software
  *  distributed under the License is distributed on an "AS IS" BASIS,
  *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,8 +15,10 @@
  *  limitations under the License.
  */
 
-library core;
+import 'package:beagle/model/request.dart';
+import 'package:beagle/model/response.dart';
 
-export 'package:beagle_core/beagle_initializer.dart';
-export 'package:beagle_core/beagle_remote_view.dart';
-export 'package:beagle_core/model/beagle_ui_element.dart';
+// ignore: one_member_abstracts
+abstract class HttpClient {
+  Future<Response> sendRequest(Request req);
+}
