+++ flutter/beagle/lib/model/beagle_ui_element.dart
