+++ flutter/beagle/lib/model/network_strategy.dart
