+++ flutter/beagle/lib/model/request.dart
