+++ flutter/beagle/lib/model/response.dart
