+++ flutter/beagle/lib/model/tree_update_mode.dart
