--- flutter/beagle_core/.metadata
@@ -1,10 +1,0 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
-
-version:
-  revision: 840c9205b344a59e48a5926ee2d791cc5640924c
-  channel: stable
-
-project_type: package
