--- flutter/beagle_core/javascript-bridge/src/global.d.ts
@@ -1,7 +1,0 @@
-type FlutterJSChannel = (
-  'httpClient.request'
-  | 'beagleView.update'
-  | 'action'
-)
-
-declare function sendMessage(channel: FlutterJSChannel, message: String)
